refactor: renamed file map.svelte.ts -> mapState.svelte.ts
